% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lookup_variable_from_excel_file.R
\name{add_lookup_variale_from_excel_file}
\alias{add_lookup_variale_from_excel_file}
\title{add_lookup_variale_from_excel_file}
\usage{
add_lookup_variale_from_excel_file(data, path_to_excel_file, match_var)
}
\arguments{
\item{data}{any data frame to add a new column from look up table}

\item{path_to_excel_file}{path to the excel file}

\item{match_var}{the column to match with in the look up table}
}
\value{
Returns dataframe with new lookup variable
}
\description{
Function that adds a new lookup variable to a dataframe from an excel file
Excelfile has two columns first the match_var column, and the second column the new column to be added.
}
