% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_lowest_value.R
\name{choose_lowest_value}
\alias{choose_lowest_value}
\title{Choose lowest value from two columns}
\usage{
choose_lowest_value(data, col1, col2, new_col_name)
}
\arguments{
\item{data}{A dataframe.}

\item{col1}{A column to compare.}

\item{col2}{A column to compare.}

\item{new_col_name}{A name for the new column.}
}
\value{
A dataframe.
}
\description{
Choose lowest value from two columns
}
\examples{
df <- data.frame(a = c(1,2,3), b = c(2, 4, 1))
choose_lowest_value(df, a, b, lowest_number)

}
