% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pontutils-package.R
\docType{package}
\name{pontutils-package}
\alias{pontutils}
\alias{pontutils-package}
\title{pontutils: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PontusChallis/pontutils}
  \item Report bugs at \url{https://github.com/PontusChallis/pontutils/issues}
}

}
\author{
\strong{Maintainer}: Pontus Challis \email{pontus.challis@umu.se} (\href{https://orcid.org/0000-0002-7511-7217}{ORCID})

}
\keyword{internal}
