% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonatal_functions.R
\name{pontutils_column_to_gw}
\alias{pontutils_column_to_gw}
\title{Convert a column of days to hole weeks}
\usage{
pontutils_column_to_gw(data, column)
}
\arguments{
\item{data}{a dataframe with the column}

\item{column}{the column in age to convert}
}
\value{
the dataframe with the new column namned {{column}}_in_gw
}
\description{
Convert a column of days to hole weeks
}
\details{
Title
}
