% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranage_of_column.R
\name{range_of_column}
\alias{range_of_column}
\title{Returns range of column}
\usage{
range_of_column(data, column)
}
\arguments{
\item{data}{A dataframe.}

\item{column}{A column to get range from.}
}
\value{
A range of a column.
}
\description{
Returns range of column
}
\examples{
df <- data.frame(a = c(1,2,3), b = c(NA, NA, NA), c = c(NA, 2, NA))
range_of_column(df, a)

}
