% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_columns.R
\name{remove_empty_columns}
\alias{remove_empty_columns}
\title{Removes empty columns from a dataframe}
\usage{
remove_empty_columns(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Removes empty columns from a dataframe
}
\examples{
df <- data.frame(a = c(1,2,3), b = c(NA, NA, NA), c = c(NA, 2, NA))
remove_empty_columns(df)
}
