% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_of_columns.R
\name{sum_of_columns}
\alias{sum_of_columns}
\title{Sum of columns}
\usage{
sum_of_columns(data, new_column, ...)
}
\arguments{
\item{data}{A dataframe.}

\item{new_column}{A name for the new column.}

\item{...}{Columns to sum.}
}
\value{
A dataframe.
}
\description{
Sum of columns
}
\examples{
df <- data.frame(a = c(1,2,3), b = c(4,2,1))
sum_of_columns(df, "sum_of_a_b", a, b)
}
